cmd/ethereum: new JS repl with cross-platform line editing
