--- cmd/ethereum/repl/repl_linux.go
@@ -1,1 +1,0 @@
-repl_darwin.go
